QATlib: 21.08.0 release
Add QATlib release 21.08.0.

+++ quickassist/lookaside/access_layer/src/sample_code/performance/common/qat_perf_buffer_utils.c
@@ -5,7 +5,7 @@
  * 
  *   GPL LICENSE SUMMARY
  * 
- *   Copyright(c) 2007-2020 Intel Corporation. All rights reserved.
+ *   Copyright(c) 2007-2021 Intel Corporation. All rights reserved.
  * 
  *   This program is free software; you can redistribute it and/or modify
  *   it under the terms of version 2 of the GNU General Public License as
@@ -27,7 +27,7 @@
  * 
  *   BSD LICENSE
  * 
- *   Copyright(c) 2007-2020 Intel Corporation. All rights reserved.
+ *   Copyright(c) 2007-2021 Intel Corporation. All rights reserved.
  *   All rights reserved.
  * 
  *   Redistribution and use in source and binary forms, with or without
@@ -601,7 +601,7 @@ CpaPhysicalAddr virtAddrToDevAddr(void *pVirtAddr,
 
     if (CPA_STATUS_SUCCESS != status)
     {
-        return (CpaPhysicalAddr)NULL;
+        return (CpaPhysicalAddr)(uintptr_t)NULL;
     }
 
     if (instanceInfo.requiresPhysicallyContiguousMemory)
@@ -610,6 +610,6 @@ CpaPhysicalAddr virtAddrToDevAddr(void *pVirtAddr,
     }
     else
     {
-        return (CpaPhysicalAddr)pVirtAddr;
+        return (CpaPhysicalAddr)(uintptr_t)pVirtAddr;
     }
 }
